Move batchRequest to contracts.js
