+++ scripts/web3.eth.contract.js
@@ -63,4 +63,28 @@ const getEvents = async () => {
   );
 };
 
-module.exports = { deploy, updateProperties, getEvents };
+const batchRequest = async () => {
+  var myContract = new web3.eth.Contract(ABI, CONTRACT_ADDRESS);
+  var batch = await new web3.BatchRequest();
+  batch.add(
+    web3.eth.getBalance.request(
+      process.env.FROM_ADDRESS,
+      'latest',
+      (error, result) => {
+        if (error) console.error(error);
+        else console.log('Result: ', web3.utils.fromWei(result, 'ether'));
+      }
+    )
+  );
+  batch.add(
+    myContract.methods
+      .greet()
+      .call.request({ from: process.env.FROM_ADDRESS }, (error, result) => {
+        if (error) console.error(error);
+        else console.log('Result: ', result);
+      })
+  );
+  batch.execute();
+};
+
+module.exports = { deploy, updateProperties, getEvents, batchRequest };
